+++ projekt/graphics.h
@@ -25,7 +25,7 @@ int numModels;
 mat4 projection_mat;
 void graphicsInitModels();
 void addModel(ArchObject * obj,char* fileName, int texture,int shader);
-void graphicsTranslation(ModelObject* m, GLuint x, GLuint y, GLuint z);
+void graphicsTranslation(ModelObject* m, GLfloat x, GLfloat y, GLfloat z);
 void graphicsDisplay(ModelObject* m, mat4 view_mat);
 
 #endif
